+++ src/features/llms/module.ts
@@ -1,5 +1,5 @@
 import defu from 'defu'
-import { createResolver, defineNuxtModule, addTypeTemplate, addServerPlugin } from '@nuxt/kit'
+import { createResolver, defineNuxtModule, addTypeTemplate, addServerPlugin, addServerHandler } from '@nuxt/kit'
 
 export default defineNuxtModule({
   meta: {
@@ -11,14 +11,49 @@ export default defineNuxtModule({
 
     addServerPlugin(resolve('runtime/server/content-llms.plugin'))
 
+    const llmsOption = (nuxt.options as unknown as { llms: { contentRawMarkdown: false | { excludeCollections: string[] } } })?.llms
+    if (llmsOption?.contentRawMarkdown !== false) {
+      addServerHandler({ route: '/raw/**:slug.md', handler: resolve('runtime/server/routes/raw/[...slug].md.get') })
+    }
+
+    nuxt.hook('modules:done', () => {
+      const contentRawMarkdown = llmsOption?.contentRawMarkdown === false
+        ? false
+        : defu(llmsOption?.contentRawMarkdown, {
+            excludeCollections: [],
+          })
+      // @ts-expect-error -- TODO: fix types
+      nuxt.options.llms ||= {}
+      // @ts-expect-error -- TODO: fix types
+      nuxt.options.llms.contentRawMarkdown = contentRawMarkdown
+
+      nuxt.options.runtimeConfig.llms ||= {}
+      // @ts-expect-error -- TODO: fix types
+      nuxt.options.runtimeConfig.llms.contentRawMarkdown = contentRawMarkdown
+    })
+
     const typeTemplate = addTypeTemplate({
       filename: 'content/llms.d.ts' as `${string}.d.ts`,
       getContents: () => {
         return `
 import type { SQLOperator, PageCollections, PageCollectionItemBase } from '@nuxt/content'
 declare module 'nuxt-llms' {
+  interface ModuleOptions {
+    contentRawMarkdown?: false | {
+      /**
+       * Whether to rewrite the LLMs.txt file to use the raw markdown endpoint
+       * @default true
+       */
+      rewriteLLMSTxt?: boolean
+      /**
+       * Whether to exclude specific collections from the raw markdown endpoint
+       * @default []
+       */
+      excludeCollections?: string[]
+    }
+  }
   interface LLMsSection {
-    contentCollection?: keyof PageCollections
+    contentCollection?: string | keyof PageCollections
     contentFilters?: Array<{
       field: string
       operator: SQLOperator
@@ -32,11 +67,10 @@ declare module 'nitropack/types' {
     'content:llms:generate:document': (event: H3Event, doc: PageCollectionItemBase, options: ModuleOptions) => void
   }
 }
-
         `
       },
       write: true,
-    }).dst
+    }, { node: true }).dst
     nuxt.options.nitro ||= {}
     nuxt.options.nitro.typescript ||= {}
     nuxt.options.nitro.typescript.tsConfig = defu(nuxt.options.nitro.typescript.tsConfig, {